requirements.txt
attrs
black
rope
mypy
returns
flask-restx
pydantic
dependency-injector
toolz
pytest
pytest-cov
